First pass at bringing in patch mode behavior
This introduces a new parameter "patching" which can be set to
"/patching_1" in the url which will enable patch playout mode for the
dynamic MPD.

During patch playout mode the MPD will contain a PatchLocation element
that points back to the server with an additional "/patch_<time>/"
reference to signify the last information given to the client.

When the client follows the PatchLocation URI it will receive a Patch
response that is applied to the existing in-memory MPD to create the
full form MPD at the original location.

Current pieces needing further investigation:
- Segment update consistency, unclear if we have overlap
- Empty segment timeline on join, periods with an empty segment timeline
should be omitted and included on the next update