Prevent xlink, segtimelineloss, and modulo with patching
There is no specification restriction that would prevent the combined
usage of patching and these options, however their implementation in the
simulator would require significant work in order to enable. To avoid
the options being used and getting invalid responses they are explicitly
detected with patching and an exception is raised. Issues can be filed
on github to ask for enablement of a scenario.

Minor fix for location production without fixed start/end times